Refactored EffectError.
